==> 01_VGC_RAWG_Exploration.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt

# Turn off warnings to ease reading
import warnings
warnings.filterwarnings('ignore')

# 0. Loading Datasets

vgc_data = pd.read_csv('Scraping/vgsales.csv', index_col = 0) # First column is the index, so we will use it as the index

rawg_data = pd.read_csv('Scraping/rawg_games_smaller.csv', parse_dates = ['released','updated'])

vgc_data.head()

rawg_data.head()

# 1. Cleaning vgc_data

There are columns we will not use:

- Sales  columns except `Total_Shipped` and `Total_Sales`.

- `Last_Update`.

- `VGC_Score`, `Critic_Score` and `User Score`, as we are going to get scores from Metacritic.



vgc_data = vgc_data.drop(['VGC_Score','Critic_Score','User_Score','NA_Sales','EU_Sales','JP_Sales','Other_Sales','Last_Update'], axis = 1)
vgc_data.head()

We are going to transform `Total_Shipped` and `Total_Sales` to numeric (do not forget that they are in the millions scale even thogh we don't multiply them by 1,000,000).

vgc_data['Total_Shipped'] = vgc_data['Total_Shipped'].str.strip('m').astype('float')
vgc_data['Total_Sales'] = vgc_data['Total_Sales'].str.strip('m').astype('float')

---
## 1.1. Release Column

Now we proceed to clean the `Release` column.

We can observe the following:

- There are NaNs
- Days are in the ~st ~nd ~rd ~th format
- Months are in 3 letter format
- Years have the last 2 digits format

sorted(vgc_data['Release'].str[-2:].dropna().unique(), reverse = True)

We can see that:
- The years belonging to the 21th century range between 2000 and 2021
- The years belonging to the 20th century range between 1970 and 1999

Now we are going to split the string into the different elements to transform them into a more intelligible format and merge them later.

months = {'Jan':'01', 'Feb':'02', 'Mar':'03', 'Apr':'04', 'May':'05', 'Jun':'06',
          'Jul':'07', 'Aug':'08', 'Sep':'09', 'Oct':'10', 'Nov':'11', 'Dec':'12'}

def date_reformat(date):
    
    """
    This function converts string dates in the dddd mm yy (e.g. '02nd Jul 98') format into string dates in the dd-mm-YYYY format ('02-07-1998').
    """
    
    output = np.nan
    
    if type(date) != float:
        
        day = date[:2]                                  # Days
        month = date[5:8]
        year = date[-2:]
        current_year = dt.datetime.today().year - 2000
        
        month = months[month]                           # Months

        if int(year) <= current_year:                    # Years

            year = '20' + year                          # 21th century

        else:

            year = '19' + year                          # 20th century
            
        output = '-'.join([day, month, year])           # Merge into a single string in the dd-mm-YYYY format

    return output

vgc_data['Release'] = vgc_data['Release'].apply(date_reformat)  

vgc_data.sample(10)

---
## 1.2. Total_Sales & Total_Shipped Columns

Now that there's nothing left to clean (in the formatting sense), let's check for NaN values to see what to do with them (as we have already seen many).

vgc_data.isna().sum()

There are so many NaNs in the dataframe we have to decide what to do with them!

- We are interested in the sales, so taking non-NaN `Total_Sales` values seems logical (other Sales columns have more NaNs, so it won't affect as much).

- Also, there are titles having NaN `Total_Sales` values but non-NaN `Total_Shipped` values.

Thus, we want to combine these two columns, and filling the NaN values.

However, to do so, first we have to check if there is any case in which both `Total_Sales` and `Total_Shipped` are non-NaN values.

We are going to fill the NaN values with zeros (we will consider 0 as not having a value in this case) and save them to a variable.

Then, we are going to perform the comparation `!= 0` to see how many non-zero values we have in each column.

### 1.2.1. Checking Columns

Also, we cast the result to int so that we can sum them to check what we wanted to discover:

- If **sum = 0** : Both columns have either 0 or NaN.
- If **sum = 1** : Only one column have non-NaN or non-zero values.
- If **sum > 1** : Both columns have non-NaN and non-zero values.

total_sales = vgc_data['Total_Sales'].fillna(0)
total_shipped = vgc_data['Total_Shipped'].fillna(0)

zero_test = (total_sales != 0).astype(int) + (total_shipped != 0).astype(int)
zero_test.value_counts()

This means that only one of the columns will hold a non-NaN or a non-zero value for the same data point.

zeros, ones = zero_test.value_counts()
zeros / (zeros + ones)

And this means that 64.45% of the dataset would have zeros or NaNs in these two columns combined. 

Thus, in order to combine both columns and filling the NaNs, we will have to do the following:

- Save both columns to separate variables filling the NaNs with 0 (we can use the variables we used above).

- Perform a column-to-column addition.

- Remove the zeros.

- Get the median value.

- Sum `Total_Sales` and `Total_Shipped`, fill NaNs with the median value we calculated and save to a new column: `Sales`.

### 1.2.2. Combining Sales-Related Columns and Filling Missing Sales Values 

combined = total_sales + total_shipped
idx_zeros = combined.index[combined <= 0].tolist()
combined = combined.drop(index = idx_zeros)
med = combined.median()

# Median value we will use to fill the NaNs.
med

vgc_data['Sales'] = vgc_data[['Total_Shipped','Total_Sales']].sum(axis = 1, min_count = 1)\
                                                             .fillna(med)

vgc_data = vgc_data.drop(['Total_Shipped','Total_Sales'], axis = 1)

vgc_data.sample(10)

---
## 1.3. Separating Video Games IPs and Saving Dataset

We will strip `vgc_data` from the IPs (`Platform` = `Series`) and save them into a separate dataframe in case we happen to use it in the future.

vgc_ips = pd.DataFrame()

ips = vgc_data[vgc_data['Platform'] == 'Series']

ips = ips.drop(['Platform'], axis = 1)
ips

ips.to_csv("Files/01_vgc_ips.csv", encoding='utf-8', index=False)

Now we drop all rows whose `Platform` column contained `Series`.

ip_idx = vgc_data.index[vgc_data['Platform'] == 'Series'].tolist()

vgc_data = vgc_data.drop(index = ip_idx)

# 2. Saving Dataset

vgc_data.to_csv("Files/01_vgc_clean.csv", encoding='utf-8', index=False)

---
# 3. Some observations on vgc_data

### No. of Platforms

vgc_data['Platform'].value_counts()

There are **78 platforms** (`All` refers to 'multiplatform')

---

### No. of Publishers

vgc_data['Publisher'].value_counts()

There are **3295 publishers** (one is Unkown)

---

### No. of Developers

vgc_data['Developer'].value_counts()

There are **8420 developers** in total.

---

# 4. Decisions yet to make

- **What do we do with the empty `Sales` values?**

    - Discard them.
    - Fill them with values -> which values?
    
    
- **What do we do with the NaN values from the `Release` column?**
    
    - If the release date is unknown it is higly likely that the title is very old (we should check them).
    
    - If it is a very old game, we shuld discard them as they will not hold much information.
    
    - If it is an IP, it will depend on what we decide to do with `All` platform.
    
    
- **What do we do with the NaN velues from the `Developer` column?**

    - Probably same as `Release` column.


- **What do we do with the games whose `Platform` column is `All`?**

    - The sum of sales in the different platforms does not usually match the `All` total sales.
    
    - We have to take into account that the RAWG dataset has platforms grouped from the beginning and that there is no `All` category.
        - Should we unpack those raws? -> it already has >500k data points so it does not seem to be the most intelligent thing to do.
        

- **How are we going to transform these categorical variables into featrues (numerical/boolan)?**

    - Platform: Its dimension is too high to do the usual One-Hot Encoding.
    
    - Publisher y Developer: There are 2 orders of magnitude above the number of platforms, thus we have to see what do we do with them to avoid the *Curse of Dimensionality*.

---

# 5. Cleaning rawg_data

## 5.1. TBA Status

Let's see how many games are in the TBA (to be announced) state column: `tba`

rawg_data[rawg_data['tba']==True]

There are 2435 titles in the TBA state. However, we can see that there are games in that stare whose `release` is **in the past**.

Let's check how many future releases are in this dataset, using Feb 28th as the delimiter for 'future', as our VGChartz dataset has been retrieved that day.

We are also removing the rows containing these titles, as they can't have sales yet and that means that they are of no use for us.

rawg_data[rawg_data['released']>='2021-02-28']

future_titles = rawg_data.index[rawg_data['released']>='2021-02-28'].tolist()
rawg_data = rawg_data.drop(index = future_titles).reset_index(drop = True)

We can see that there are far more games that should have the `tba` but they don't. Thus, we can drop the `tba` column as it gives non-accurate information.

## 5.2. Dropping Useless Columns

Furthermore:

- We will also drop `id` as it's only useful if we are going to do requests to the API, but we are not allowed to do more than 20.000 requests a month, so nothing we can do with a dataset with >500k ids. 

- `updated` is a useless column, as our goal is not to evaluate the updates for a games' success (if we want to predict if a game is a success, it shouldn't even have an update).

- We will keep `slug` even though it seems useless, as the Metacritic page for games follows the `https://www.metacritic.com/game/PLATFORM/SLUG` format.

rawg_data = rawg_data.drop(['id','tba','updated', 'metacritic'], axis = 1)
rawg_data.head()

rawg_data.shape[0]

# 6. Some observations on rawg_data

### No. of non-NaN values

rawg_data.notna().sum()

We can see that there are very few `metacritic` values available, as well as kind of few values on `esrb_rating`.

We may opt to drop this column (`metacritic`) after retrieving the scores from the Metacritic web using a scraper.

---

### No. of NaN values

rawg_data.isna().sum()

---

We are dropping the 2 rows with `slug` and `name` missing, as we won't know which games they are.

rawg_data[rawg_data['slug'].isna()]

rawg_data = rawg_data.drop(index = [34946, 467110])

rawg_data.shape[0]

# 7. Saving Dataset

rawg_data.to_csv("Files/01_rawg_clean.csv", encoding='utf-8', index=False)

---
# 8. Decisions yet to make
---

- How are we going to merge this dataset with vgc_data?

    - We are going to lowercase `name` and `title`, and if they match, they are added to the same row.
    
        - There are problems with the `platforms`, `genres`, `stores`, and `tags` columns.
        
        - Especially `platforms`, as vgc_data is separated into different platforms.
        

- How are we going to unpack the list-like columns? Are we going to use One-Hot Encoding?

    - We still **DON'T KNOW**


- What are we going to do with NaN values in some columns after the merge?

    - We still **DON'T KNOW**
