==> 02_RAWG_Processing.ipynb <==
import pandas as pd
import numpy as np
from ast import literal_eval

import warnings
warnings.filterwarnings('ignore')

# 0. Loading Dataset

rawg_data = pd.read_csv('Files/01_rawg_clean.csv', parse_dates = ['released'])

rawg_data.head()

# 1. What We Want to Do

We want to unpack values inside the list-like columns, specially `platforms` (we can leave the rest for later, as we can directly do it on the feature engineering process).

The idea is to copy the raw for every platform listed in the column so that it looks like the following:

| slug | name | released | metacritic | suggestions_count | platforms | (...) |
| ---- | ---- | ---- | ---- | ---- | ---- | ---- |
| grand-theft-auto-v | Grand Theft Auto V | 2013-09-17 | 97.0 |	416 | PC | (...) |
| grand-theft-auto-v | Grand Theft Auto V | 2013-09-17 | 97.0 |	416 | Xbox Series S/X | (...) |
| grand-theft-auto-v | Grand Theft Auto V | 2013-09-17 | 97.0 |	416 | PlayStation 5 | (...) |


# 2. But first...

CSV files has the problem that saves lists as str, so we have to convert them back to be able to access the data inside them. To do so, we will apply the `literal_eval` function from the `ast` library.

After the first cleaning, there are rows in the list-like columns which are not lists, as the list was a single-element list. We will not apply `literal_eval` on them as they raise an error.

rawg_data['platforms'] = rawg_data['platforms'].apply(lambda x: literal_eval(x) if pd.notnull(x) and ('[' in x) else x)
rawg_data['genres'] = rawg_data['genres'].apply(lambda x: literal_eval(x) if pd.notnull(x) and ('[' in x) else x)
rawg_data['stores'] = rawg_data['stores'].apply(lambda x: literal_eval(x) if pd.notnull(x) and ('[' in x) else x)
rawg_data['tags'] = rawg_data['tags'].apply(lambda x: literal_eval(x) if pd.notnull(x) and ('[' in x) else x)

---

# 3. Checking for NaN values

Now that lists are lists again, let's look at the `NaN` values in this dataset to study them.

rawg_data.isna().sum()

### 3.1. Platforms' NaNs


On a first glance, we may think that it would be a good idea to remove them in the `platforms` column, as not having a platform means that the game won't be able to be merged into the VCG dataset.

Let's see what kind of games have `NaN` in the `platforms` column.

rawg_data[rawg_data['platforms'].isna()]

What we can see here is that when a game has a `NaN` on the `platforms` column, some other list-like columns can also have `NaN` on them.

However, if we look at, for example, the `stores` column more in detail...

rawg_data[rawg_data['platforms'].isna()]['stores'].value_counts()

We see that there are some portential problems there:

- A game that can be bought in steam should have `PC` on the `Platforms` column.

- The same happens to `Playstation Store` and `Xbox Store`. The problem here is that they are DLCs as we can see below, and they already have their counterpart for consoles on the dataset.

nan_plat = rawg_data[rawg_data['platforms'].isna()]
ps_store = rawg_data['stores'] == 'PlayStation Store'

nan_plat[ps_store]

Let's look for some samples from here:

rawg_data[rawg_data['name'].apply(lambda x: False if type(x) == float or 'Assassin\'s Creed Valhalla' not in x else True)] # NaNs didn't let us filter normally

rawg_data[rawg_data['name'].apply(lambda x: False if type(x) == float or 'Borderlands 3' not in x else True)] # NaNs didn't let us filter normally

Let's see if we can gather more insight on the `Steam` department.

steam = rawg_data['stores'] == 'Steam'

nan_plat[steam]

What we see is that **2 of them are MODs (= free content)** and the other 2 are games **not released yet**. Thus, we can remove all these data without losing information. as we do not want mobile-exclusive games and we will not consider itch.io for the moment, as the VGC dataset does not contain data from there, thus no sales data.

---

## 3.2. Dropping Rows

Let's then proceed with the removal of rows having the one or more of the next characteristics:

- `platforms` is `NaN`

- Mobile-only games, thus `platforms` only contain whether `Android` or `iOS`

- `stores` has only `itch.io`


plat_nan = rawg_data.index[rawg_data['platforms'].isna()].tolist()
android = rawg_data.index[rawg_data['platforms']=='Android'].tolist()
ios = rawg_data.index[rawg_data['platforms']=='iOS'].tolist()

plat_to_drop = plat_nan + android + ios

rawg_data = rawg_data.drop(index=plat_to_drop)\
                     .reset_index(drop=True)

rawg_data.shape[0]

itchio = rawg_data.index[rawg_data['stores']=='itch.io'].tolist()

rawg_data = rawg_data.drop(index=itchio)\
                     .reset_index(drop=True)

rawg_data.shape[0]

We have drastically reduced the dimension of the dataset to almost 20% of the original size.

Let's check how many `NaN` do we have:

rawg_data.isna().sum()

nan_release = rawg_data.index[rawg_data['released'].isna()].tolist()
                   
rawg_data = rawg_data.drop(index = nan_release)\
                     .reset_index(drop=True)

rawg_data

---

# 4. Unpacking Columns

Now we would want to unpack the rows to single-platform columns...but how do we accomplish that?

We could do the following:

- Create an auxiliar DataFrame `rawg_aux`

- Iterate over every row with `for idx, row in rawg_data.iterrows()`

- For every platform, append the columns to `rawg_aux`

\* **Ideally**, we would to do this step with a function being applied to the dataframe. 

However, we haven't come up with the idea to how to do it yet.

rawg_aux = pd.DataFrame(columns = rawg_data.columns)

### **Do not execute the code below as it takes a lot to finish executing**

for idx, row in rawg_data.iterrows():
    
    plats = row['platforms']
    
    if type(plats) != str: # We do not want to do this loop if the game only has one platform (thus, no list-like = string).
        for element in plats:      

            aux = row.copy()
            aux['platforms'] = element

            rawg_aux = rawg_aux.append(aux)
    
    else:
        
        aux = row.copy()
        aux['platforms'] = plats

        rawg_aux = rawg_aux.append(aux)
        
    if idx % 10000 == 0:
        print(idx)


rawg_aux = rawg_aux.rename(columns = {'platforms': 'platform',
                                      'esrb_rating': 'esrb'})

print('Finished!')

rawg_data = rawg_aux.reset_index(drop=True)
rawg_data

---
# 5. Checking For Mobile Games After Unpacking

sorted(rawg_data['platform'].unique())

We can see that after unpacking the `platforms` column into `platform`, we get `iOS` and `Android` again.

We are going to drop the raws containing those data by getting the indexes and using the `pd.DataFrame.drop()` method.

ios = rawg_data.index[rawg_data['platform']=='iOS'].tolist()
android = rawg_data.index[rawg_data['platform']=='Android'].tolist()

drop = ios + android

rawg_data = rawg_data.drop(index = drop)\
                     .reset_index(drop=True)

rawg_data

# 6. Getting URLS Scrape Metacritic's Web Site

## 6.1. Listing the platforms available in Metacritic into a dictionary

Remember the metacritic page for a game has the next format:

**https://www.metacritic.com/game/PLATFORM/SLUG**

But not every platform in this dataset is available, as this page did not exist back then.

We will list the platforms available in Metacritic inside a dictionary, using the `Platform` column counterpart as the key value:
`RAWG-platform: Metacritic-platform`

platforms_metacritic = {
    
    # Sony
    
    "PlayStation 5": "playstation-5",
    "PlayStation 4": "playstation-4",
    "PlayStation 3": "ps3",
    "PlayStation 2": "ps2",
    "PlayStation": "ps",
    "PS Vita": "vita",
    "PSP": "psp",
    
    # Microsoft
    
    "Xbox One": "xbox-one",
    "Xbox Series S/X": "xbox-series-x",
    "Xbox 360": "xbox-360",
    "Xbox": "xbox",
    
    # Nintendo
    
    "Nintendo Switch": "switch",
    "Wii U": "wii-u",
    "Wii": "wii",
    "GameCube": "gamecube",
    "Nintendo 64": "n64",
    "Nintendo 3DS": "3ds",
    "Nintendo DS": "ds",
    "Nintendo DSi": "ds",
    "Game Boy Advance": "gba",
    
    # Others
    
    "PC": "pc",
    "Dreamcast": "dreamcast"
    
}

---

We are going to add a new column `plat_mc` to `rawg_data` for the platform in Metacritic format. Any console not in the `platforms_metacritic` dictionary will have a `NaN` assigned.

def plat_metacritic(x):
    
    '''
    Given a string indicating a platform, returns the equivalent for Metacritic.
    
    x: String indicating a platform.
    '''
    
    if x in platforms_metacritic.keys():
        
        result = platforms_metacritic[x]
        
    else:
        
        result = np.nan
        
    return result

rawg_data['plat_mc'] = rawg_data['platform'].apply(plat_metacritic)

rawg_data

rawg_data['plat_mc'].value_counts()

Indexes seem to have set all to 0 somehow, so we will reset them with the `pd.DataFrame.reset_index()` method.

rawg_data.head()

rawg_data = rawg_data.reset_index(drop=True)

We will list the indexes of NaN in the `plat_mc` column in order to drop them.

no_metacritic = rawg_data.index[rawg_data['plat_mc'].isna()].tolist()

rawg_data = rawg_data.drop(index = no_metacritic)\
                     .reset_index(drop = True)

# 7. Creating URL Column

We are creating a new column named `url`, which will include the hyperlink to the games' Metacritic page (if it exists). This will ease the web scraping process.

rawg_data['url'] = 'https://www.metacritic.com/game/' + rawg_data['plat_mc'] + '/' + rawg_data['slug']

rawg_data.head()

# 8. Saving Dataset

rawg_data.to_csv("Files/02_rawg_metacritic_url.csv", encoding='utf-8', index=False)
