==> 03_Metacritic_Scraping.ipynb <==
from bs4 import BeautifulSoup, element
import requests
import pandas as pd
import numpy as np

# 0. Loading Dataset

Let's load the RAWG dataset containing urls for the Metacritic Web Site

rawg_data = pd.read_csv('Files/02_rawg_metacritic_url.csv', parse_dates = ['released'])

rawg_data.head()

# 1. Scraping Metacritic

user_agent = {'User-agent': 'Mozilla/5.0'}

urls =[]

c_score = []
c_reviews = []
c_positive = []
c_mixed = []
c_negative = []

u_score = []
u_reviews = []
u_positive = []
u_mixed = []
u_negative = []

### **The code below takes ~4 days to finish running DO NOT EXECUTE FOR NOW**

### ~4 days of expected execution

for row in rawg_data['url']:

    ##########################
    ##### Critic Reviews #####
    ##########################
    
    url = row + '/critic-reviews'
    
    response = requests.get(url, headers = user_agent)
    soup = BeautifulSoup(response.text, 'html.parser')
    
        ### Variables ###  ------- Default values in case conditions are not met
        
    critics_score = np.nan
    critics_reviews = np.nan
    critics_positive = np.nan
    critics_mixed = np.nan
    critics_negative = np.nan
    
        ### Conditions ###
        
    page_exist = soup.find('div', class_ = 'module errorcode_module error404_module') == None # Game does have a Metacritic page
    has_rating = soup.find('span', itemprop = 'ratingValue') != None
    has_score_distrib = soup.find('div', class_="score_distribution") != None
    
        ### Main process ###
    
    try:
        if  page_exist: 

            if has_rating and has_score_distrib:
                critics_score = soup.find('span', itemprop = 'ratingValue').string.strip()
                critics_reviews = soup.find('span', class_="count").find('strong').string.strip().split()[0]
                critics_positive = soup.find('div', class_="score_distribution").find_all('span', class_="count")[0].string.strip()
                critics_mixed = soup.find('div', class_="score_distribution").find_all('span', class_="count")[1].string.strip()
                critics_negative = soup.find('div', class_="score_distribution").find_all('span', class_="count")[2].string.strip()
                
    
    except AttributeError: 
        print('Error for critics here: ' + url)

    
    ########################
    ##### User Reviews #####
    ########################
    
    url = row + '/user-reviews'
    
    response = requests.get(url, headers = user_agent)
    soup = BeautifulSoup(response.text, 'html.parser')
    
        ### Variables ###  ------- Default values in case conditions are not met
        
    user_score = np.nan
    user_reviews = np.nan
    user_positive = np.nan
    user_mixed = np.nan
    user_negative = np.nan
    
        ### Conditions ###
    
    page_exist = soup.find('div', class_ = 'module errorcode_module error404_module') == None # Game does have a Metacritic page
    has_rating = soup.find('div', class_ = 'userscore_wrap feature_userscore') != None
    has_score_distrib = soup.find('div', class_="score_distribution") != None
    
        ### Main process ###
        
    if  page_exist: 
        
        try:
            if has_rating and has_score_distrib:
                user_score = soup.find('div', class_ = 'score_summary').text.split('\n')[3]
                try:
                    user_reviews = soup.find('span', class_="count").find('strong').string.strip().split()[0]
                except:
                    print('Not enough reviews')
                user_positive = soup.find('div', class_="score_distribution").find_all('span', class_="count")[0].string.strip()
                user_mixed = soup.find('div', class_="score_distribution").find_all('span', class_="count")[1].string.strip()
                user_negative = soup.find('div', class_="score_distribution").find_all('span', class_="count")[2].string.strip()
            
        except AttributeError:
            print('Error for user here: ' + url)
    
    ##############################
    ##### Updating the lists #####
    ##############################
    
    urls.append(row)
    
    c_score.append(critics_score)
    c_reviews.append(critics_reviews)
    c_positive.append(critics_positive)
    c_mixed.append(critics_mixed)
    c_negative.append(critics_negative)

    u_score.append(user_score)
    u_reviews.append(user_reviews)
    u_positive.append(user_positive)
    u_mixed.append(user_mixed)
    u_negative.append(user_negative)

        
print('Finished!')

# 2. Conversion into DataFrame

We convert the lists into a DataFrame so that we can save it as a `.csv` file.

ratings = pd.DataFrame.from_dict({'url': urls,
                                  'critic_score': c_score,
                                  'critic_reviews': c_reviews,
                                  'critic_pos': c_positive,
                                  'critic_mix': c_mixed,
                                  'critic_neg': c_negative,
                                  'user_score': u_score,
                                  'user_reviews': u_reviews,
                                  'user_pos': u_positive,
                                  'user_mix': u_mixed,
                                  'user_neg': u_negative})

ratings.shape[0]

ratings.isna().sum()

# 3. Saving Dataset

ratings.to_csv("Files/03_metacritic.csv", encoding='utf-8', index=False)
