==> 04_VGC_RAWG_Metacritic_Processing.ipynb <==
import pandas as pd
import numpy as np
from ast import literal_eval

# Turn off warnings to ease reading
import warnings
warnings.filterwarnings('ignore')

# 0. Loading Datasets

vgc_data = pd.read_csv('Files/01_vgc_clean.csv', parse_dates = ['Release'])
rawg_data = pd.read_csv('Files/02_rawg_metacritic_url.csv', parse_dates = ['released'])
metacritic = pd.read_csv('Files/03_metacritic.csv')


### Converting lists interpreted as strings to lists ###

rawg_data['genres'] = rawg_data['genres'].apply(lambda x: literal_eval(x) if pd.notnull(x) and ('[' in x) else x)
rawg_data['stores'] = rawg_data['stores'].apply(lambda x: literal_eval(x) if pd.notnull(x) and ('[' in x) else x)
rawg_data['tags'] = rawg_data['tags'].apply(lambda x: literal_eval(x) if pd.notnull(x) and ('[' in x) else x)

vgc_data.head()

rawg_data.head()

metacritic.head()

# 1. Removing Columns

We no longer need the next columns in `rawg_data`:

- `slug` : We only needed this column to create the url.

- `released` : We already have this column in the main `vgc_data` dataframe in the correct format (DD-MM-YYYY).

- `metacritic` : We have an up-to-date dataset with all this information.

- `plat_mc` : We only needed this column to create the url.

rawg = rawg_data.drop(['slug','released', 'plat_mc'], axis = 1)

rawg.head()

# 2. Merging RAWG and Metacritic Datasets

We will do this by joining both dataframes by the `pd.DataFrame.join()` method, using the `url` columns as indexes.

rawg = rawg.set_index('url')
metacritic = metacritic.set_index('url')

rawg_meta = rawg.join(metacritic)
rawg_meta.head()

We no longer need `url`, so we drop it using `reset_index` as we converted it to our index.

rawg_meta = rawg_meta.reset_index(drop = True)
rawg_meta.head()

---
# 3. Merging RAWG and VGC Datasets
## 3.1. Platforms in `rawg_meta` VS `vgc_data` Datasets

In order to merge both datasets, the columns giving the information about the platform have to have the same format. We already know the column from `rawg_meta`. Let's see what `vgc_data` has:

vgc_data['Platform'].unique()

---

We can see that it doesn't match the `rawg_meta` format at all, which is:

rawg_meta['platform'].unique()

---
Let's first drop the undesired platforms (iOS, Android and Browser, as they are usually free and we don't want to study them in the same category as console games).

After that, we will filter out those games with sales < 0 and check if the unique platforms list gets reduced or not.

### 3.2. Dropping rows containing undesired platforms (free platforms)

# Check shape before dropping
vgc_data.shape[0]

ios = vgc_data.index[vgc_data['Platform']=='iOS'].tolist()
android = vgc_data.index[vgc_data['Platform']=='And'].tolist()
browser = vgc_data.index[vgc_data['Platform']=='BRW'].tolist()
undesired = ios + android + browser

vgc_data = vgc_data.drop(index = undesired)\
                   .reset_index(drop = True)

# Check shape after dropping
vgc_data.shape[0]

We reduced the dimension of the dataframe by a huge amount (~36% of the original).

vgc_data['Platform'].unique()

We can see that the number of platforms has also been reduced.

## 3.3. Checking Consoles

Let's discover now which console is which in this dataframe:

- **All**: All consoles available for the title.
- **Wii**: Nintendo Wii
- **NES**: Nintendo Entertainment System
- **PC**: PC
- **NS**: Nintendo Switch
- **GB**: Nintendo GameBoy
- **DS**: Nintendo DS
- **X360**: Xbox 360
- **SNES**: Super Nintendo Entertainment System
- **PS3**: PlayStation 3
- **PS4**: PlayStation 4
- **3DS**: Nintendo 3DS
- **PS2**: PlayStation 2
- **GBA**: Nintendo GameBoy Advance
- **GEN**: Sega Genesis
- **N64**: Nintendo 64
- **PS**: PlayStation
- **XOne**: Xbox One
- **WiiU**: Nintendo Wii U
- **XB**: Xbox
- **PSP**: PlayStation Portable
- **2600**: Atari 2600
- **GC**: Nintendo GameCube
- **GBC**: Nintendo GameBoy Color
- **PSN**: PlayStation Network
- **PSV**: PlayStation Vita
- **DC**: Dreamcast
- **XBL**: Xbox Live
- **SAT**: Sega Saturn
- **SCD**: Sega CD System
- **3DO**: 3DO Interactive Multiplayer
- **WS**: Wonder Swan
- **Amig**: Commodore Amiga
- **VC**: Nintendo Virtual Console
- **NG**: Neo Geo
- **WW**: Nintendo WiiWare (online shop) --> **Count as Wii**
- **PCE**: PC Engine
- **GG**: Game Gear
- **OSX**: Apple OSX --> **Count as PC**
- **PCFX**: PC-FX
- **Mob**: Mini Mobile

This list, brought into a table would look like this:

| Multiplatform | PC | Sony | MicroSoft | Nintendo | Others |
|---------------|----|------|-----------|----------|--------|
| All | PC<br>OSX | PlayStation 4<br>PlayStation 3<br>PlayStation 2<br>PlayStation<br>PlayStation Network<br>PlayStation Portable<br>PlayStation Vita | Xbox One<br>Xbox 360<br>Xbox<br>Xbox Live | Switch<br>Wii U<br>Wii<br>GameCube<br>Nintendo 64<br>SNES<br>NES<br>Virtual Console<br>3DS<br>DS<br>GBA<br>GBC<br>GB | Sega Genesis<br>Sega Saturn<br>Sega CD System<br>Atari 2600<br>Dreamcast<br>3DO Interactive Multiplayer<br>Wonder Swan<br>Commodore Amiga<br>Neo Geo<br>PC Engine<br>Game Gear<br>PC-FX<br>Mini Mobile |


## 3.4. Transforming Consoles

In order to merge both datasets, we should create a new column called `join_platform` and assigning values on the following dictionaries using each dataset's platforms column as keys:

# VGC

platforms_vgc = {
    
    # All
    
    "All": "All",
    
    # PC
    
    "PC": "PC",
    "OSX": "PC",
    
    # Sony

    "PS4": "PS4",
    "PS3": "PS3",
    "PS2": "PS2",
    "PS": "PS",
    "PSN": "PSN",
    "PSV": "Vita",
    "PSP": "PSP",
    
    # Microsoft
    
    "XOne": "XOne",
    "X360": "X360",
    "XB": "Xbox",
    "XBL": "Xbox Live",
    
    # Nintendo
    
    "NS": "Switch",
    "WiiU": "Wii U",
    "Wii": "Wii",
    "WW": "Wii",
    "VC": "VC",
    "GC": "GCube",
    "N64": "N64",
    "SNES": "SNES",
    "NES": "NES",
    "3DS": "3DS",
    "DS": "DS",
    "GBA": "GBA",
    "GBC": "GBC",
    "GB": "GB",
    
    # Others
    
    "DC": "Dreamcast",
    "GEN": "Genesis",
    "2600": "Atari 2600",
    "SAT": "Saturn",
    "SCD": "Sega CD",
    "3DO": "3DO",
    "WS": "WSwan",
    "Amig": "Amiga",
    "NG": "NeoGeo",
    "PCE": "PC Engine",
    "GG": "Game Gear",
    "PCFX": "PCFX",
    "Mob": "Mini Mobile"
    
}

#RAWG_Meta

platforms_rawg = {
    
    # Sony
    
    "PlayStation 5": "PS5",
    "PlayStation 4": "PS4",
    "PlayStation 3": "PS3",
    "PlayStation 2": "PS2",
    "PlayStation": "PS",
    "PS Vita": "Vita",
    "PSP": "PSP",
    
    # Microsoft
    
    "Xbox One": "XOne",
    "Xbox Series S/X": "XS/X",
    "Xbox 360": "X360",
    "Xbox": "Xbox",
    
    # Nintendo
    
    "Nintendo Switch": "Switch",
    "Wii U": "Wii U",
    "Wii": "Wii",
    "GameCube": "GCube",
    "Nintendo 64": "N64",
    "Nintendo 3DS": "3DS",
    "Nintendo DS": "DS",
    "Nintendo DSi": "DS",
    "Game Boy Advance": "GBA",
    
    # Others
    
    "PC": "PC",
    "Dreamcast": "Dreamcast"
    
}

vgc_data['Platform'] = vgc_data['Platform'].map(platforms_vgc)
rawg_meta['platform'] = rawg_meta['platform'].map(platforms_rawg)

vgc_data

rawg_meta

---
## 3.5. Merging

We want to join both dataframes taking these into account:

- Game title
- Platform

But if we did a regular join, this would be the result:

vgc_data.set_index('Title').join(rawg_meta.set_index('name'))

We can see that `Platform` and `platform` do not match each other (and thus, now the dataset is bigger than it should be).

To solve this problem, we will lowercase the game title and add the platform string to it:

rawg_meta['join_col'] = rawg_meta['name'] + ' ' + rawg_meta['platform']
rawg_meta = rawg_meta.set_index('join_col')

vgc_data['join_col'] = vgc_data['Title'] + ' ' + vgc_data['Platform']
vgc_data = vgc_data.set_index('join_col')

We drop the `join_col` indexes and the `name` and `platform` columns as we join both dataframes.

data = vgc_data.join(rawg_meta).reset_index(drop = True).drop(['name','platform'], axis = 1)
data

## 3.6. Renaming Columns
We rename the columns from the `rawg_meta` dataframe to match the column names from the VGC dataframe.

rename = {
    "suggestions_count": "Suggest_count",
    "genres": "Genres",
    "stores": "Stores",
    "tags": "Tags",
    "esrb": "ESRB",
    "critic_score": "C_Score",
    "critic_reviews": "C_Reviews",
    "critic_pos": "C_Positive",
    "critic_mix": "C_Mixed",
    "critic_neg": "C_Negative",
    "user_score": "U_Score",
    "user_reviews": "U_Reviews",
    "user_pos": "U_Positive",
    "user_mix": "U_Mixed",
    "user_neg": "U_Negative"
}

data = data.rename(columns = rename)
data.head()

# 4. Checking Values in Critic and User-Related Columns 

check_c = data.filter(like = 'C_', axis = 1)
check_u = data.filter(like = 'U_', axis = 1)

check_c.dtypes

check_u.dtypes

We see that the **User-related columns have strings in them**. We must transform them to numeric.

- `U_Score` seems to have some of their values as `tbd`, so we will transform them to NaN.
- The rest of user columns contain commas, so we are going to replace them with a blank space.

Then we are filtering the dataframe to apply a `pd.to_numeric` transformation to all of them.

# We replace 'tbd' with NaN
data['U_Score'] = data['U_Score'].map(lambda x: x if x != 'tbd' else np.nan)

# We remove commas from thousands
data['U_Positive'] = data['U_Positive'].str.replace(',','')
data['U_Mixed'] = data['U_Mixed'].str.replace(',','')
data['U_Negative'] = data['U_Negative'].str.replace(',','')

data[data.filter(like = 'U_', axis = 1).columns] = data.filter(like = 'U_', axis = 1).apply(pd.to_numeric)

data['U_Score'] = data['U_Score'] * 10 # We multiply this column by 10 so that it is in the same scale as C_Score (out of 100 instead of 10).

We check if the transformation was done correctly.

data.filter(like = 'U_', axis = 1).dtypes

# 5. Saving Dataset

data.to_csv("Files/04_merged_data.csv", encoding='utf-8', index=False)

---
# 6. Dealing with some features

Some columns with categorical values such as:

- `Publisher`

- `Developer`

- `Platform`

- `Genres`

- `Stores`

- `Tags`

are potentially a problem, as applying One-Hot Encoding to each of them would cause **problems in dimensionality**, mainly due to `Tags` containing **more than 1000 unique values**.

## 6.1. Counting Unique Values

Let's quicky check how many unique values they contain:

def unique_nested_values(a_col, a_list):
    
    '''
    Given a pd.Series object and a list, retuns the count of unique values inside the Series object in the input list.
    
    a_col: Column of interest.
    a_list: List where we want the counts to be stored on.
    '''
    
    col = a_col
    result = a_list
    
    if type(col) != list:
        if col not in result:
            result.append(col)
    else:
        for element in col:
            if element not in result:
                result.append(element)
    
    return result

genres = []
stores = []
tags = []

for idx, row in data.iterrows():
    
    genre = row['Genres']
    store = row['Stores']
    tag = row['Tags']

    unique_nested_values(genre, genres)
    unique_nested_values(store, stores)
    unique_nested_values(tag, tags)

print(str(data['Platform'].value_counts().shape[0]) + ' unique values')
data['Platform'].value_counts()

print(str(data['Publisher'].value_counts().shape[0]) + ' unique values')
data['Publisher'].value_counts()

---

print(str(data['Developer'].value_counts().shape[0]) + ' unique values')
data['Developer'].value_counts()

---

print(str(len(genres)) + ' unique values')
pd.Series(genres)

---

print(str(len(stores)) + ' unique values')
pd.Series(stores)

---

print(str(len(tags)) + ' unique values')
pd.Series(tags)

---

So we have:

|**Column** |**Unique Values**|**Comments**                                  |
|-----------|-----------------|----------------------------------------------|
|*Platform* |39               |                                              |
|*Publisher*|3053             |Will cause dimensionality problem!            |
|*Developer*|8371             |Will cause dimensionality problem!            |
|*Genres*   |20               |                                              |
|*Stores*   |11               |We should remove *App Store* and *Google Play*|
|*Tags*     |1379             |Will cause dimensionality problem!            |


As we can see, having a dataframe with >1500 columns does not look like a good idea. Thus, we have to take some decisions regarding these columns!

## 6.2. What to do with categorical columns

- What do we do with the `Platform` column?
    
    - Remove `All`s, as they take every platform for the title available at the moment (and sometimes, the sum does not match the figures in here).
    
    - Convert the platforms into the company selling the platform (e.g. 'Playstation 4' -> 'Sony', 'Wii' -> 'Nintendo'), summing the sales for the same company platform.
    
    
- What do we do with the `Genres` column?

    - As there are not many unique values, One-Hot Encoding is an option.
    
 
- What do we do with the `Stores` column?

    - Same as the `Genres` column.


- What do we do with the `Tags` column?

    - We have to research what we can do in this case.
    
    - One option found is using the top-10 most frequent tags for One-Hot Encoding.
    
    - Not using tags for the analysis.


- What do we do with the `Publisher` column?

    - Same thing as `Tags` column.
    
    - Target encoding the top 30~50 values.


- What do we do with the `Developer` column?

    - Same thing as `Tags` column.
    
    - Target encoding the top 30~50 values.
